Reset some resultType's
Manually setting `resultType` for the `ClassOrModule` symbols to
`nullptr` and to a `LambdaParam(tp, todo, todo)` for the `AttachedClass`
type template mimics what would be the state had we just called
`insertClass` for the first time, and removes all references of the
newly-deleted type members from the containing class.

This forces resolver to recompute these types, taking into account any
deleted and added type members.